fix: 29 bug fixes + AVX-512 performance + GGUF v3 dtype 30 support (#1339)
* ponytail: purge ~48K lines of dead code

192 files deleted — engines, tools, tests, kernels, headers never compiled or called.
43% codebase mass removed. Build verified clean. Zero behavior change.

Includes bug fixes from prior sweep:
- SIGPIPE handler in server
- double-free prevention (backend_hip, backend_cuda, backend_vulkan)
- fopen null checks in npu_engine_gguf
- ModelData leak fix
- prefill_dispatcher UB fix
- model_discovery dead code fix
- command injection hardening
- rope cache data race fix
- cached_dlopen data race fix
- volatile→atomic in backend_demo
- signal handler safety fix
- getenv HOME null checks

* fix: 29 bug fixes + AVX-512 perf + GGUF v3 dtype 30 support

Bug fixes:
- SIGPIPE handler in server
- double-free prevention (hip/cuda/vulkan backends)
- fopen null checks, ModelData leak, prefill UB shift
- H1B magic check, Q4NX/safetensors overflow bypass
- ONNX varint UB, path traversal hardening
- NPU binary validation, cached_dlopen race
- rope cache data race, extern C exception safety
- TensorDesc ABI, colibri AVX-512 header fix
- JSON type validation, billing atof, config crash
- DynamicRouter lock, GPU sync removal
- getenv HOME null guards, fscanf validation
- GGUF v3 dtype 30 (F32) loading support

Performance:
- AVX-512 matmul (16-wide FMA intrinsics)
- AVX-512 SiLU (vectorized tanh approximation)
- -march=native -ffast-math on 7 targets
- 14 per-token heap allocs -> pre-allocated members
- getenv cache, RoPE freq table, attention scale hoist
- KV cache index stride precomputed
- NPU paths absolute (not ./ relative)
- /tmp fallback removed

25 files, 297 insertions, 133 deletions

---------

+++ src/zamba2_engine.h
@@ -223,7 +223,7 @@ bool load_zamba2_from_gguf(const std::string& path, Zamba2Model& model);
 
 // ── Utility: RMS Norm ──
 inline void rms_norm(const float* x, float* y, const float* w, int n, float eps) {
-    float ss = 0.0f;
+    double ss = 0.0;  // double accumulation for precision (fixes #1320)
     for (int i = 0; i < n; ++i) ss += x[i] * x[i];
     float rms = std::sqrt(ss / n + eps);
     float inv_rms = 1.0f / rms;
